+++ source/reconstructions_main.m
@@ -942,7 +942,7 @@ else
 end
 
 if subsets > 1
-    pituus = [int64(0);cumsum(pituus)];
+    pituus = [int64(0);int64(cumsum(pituus))];
     if iscell(index)
         index = cell2mat(index);
     end
@@ -1449,9 +1449,9 @@ else
                                 TOFSize = int64(size(xy_index_input,1));
                             end
                             if options.projector_type == 2 || options.projector_type == 3
-                                lor2 = [uint64(0); cumsum(uint64(lor_orth(pituus(osa_iter)+1:pituus(osa_iter + 1))))];
+                                lor2 = [uint64(0); uint64(cumsum(lor_orth(pituus(osa_iter)+1:pituus(osa_iter + 1))))];
                             else
-                                lor2 = [uint64(0); cumsum(uint64(lor_a(pituus(osa_iter)+1:pituus(osa_iter + 1))))];
+                                lor2 = [uint64(0); uint64(cumsum(lor_a(pituus(osa_iter)+1:pituus(osa_iter + 1))))];
                             end
                             if exist('OCTAVE_VERSION','builtin') == 0
                                 [A, ll] = projector_mex( Ny, Nx, Nz, dx, dz, by, bx, bz, z_det, x, y, dy, yy, xx , NSinos, NSlices, size_x, zmax, options.vaimennus, ...
